ref pojobasedpartition
